requirements.txt
fastapi==0.111.0
pymodbus==3.6.6
uvicorn==0.30.1
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3